# quicvarint: speed up parsing of 1, 2 and 4 byte varints 