--- a/quicvarint/varint.go
+++ b/quicvarint/varint.go
@@ -1,6 +1,7 @@
 package quicvarint
 
 import (
+	"encoding/binary"
 	"fmt"
 	"io"
 )
@@ -92,18 +93,13 @@ func Parse(b []byte) (uint64 /* value */, int /* bytes consumed */, error) {
 			uint64(b[1])<<16 |
 			uint64(b[2])<<8 |
 			uint64(b[3]), 4, nil
-	case 3: // 8-byte encoding: 11xxxxxx
+	case 3: // 8-byte encoding
 		if len(b) < 8 {
 			return 0, 0, io.ErrUnexpectedEOF
 		}
-		return uint64(first&0x3F)<<56 |
-			uint64(b[1])<<48 |
-			uint64(b[2])<<40 |
-			uint64(b[3])<<32 |
-			uint64(b[4])<<24 |
-			uint64(b[5])<<16 |
-			uint64(b[6])<<8 |
-			uint64(b[7]), 8, nil
+		tmp := binary.BigEndian.Uint64(b[0:8])
+		tmp &= 0x3FFFFFFFFFFFFFFF // clear top 2 bits
+		return tmp, 8, nil
 	}
 
 	// Should never be reached