@marten-seemann 2025-06-24
Here's the benchmark of the change that we merged:
```
name                     old time/op  new time/op  delta
Read/1-byte-16           3.32ns ± 3%  3.27ns ± 2%   -1.53%  (p=0.000 n=20+20)
Read/2-byte-16           4.81ns ± 3%  4.86ns ± 2%   +0.88%  (p=0.032 n=20+20)
Read/4-byte-16           7.22ns ± 2%  7.44ns ± 1%   +2.95%  (p=0.000 n=18+20)
Read/8-byte-16           12.6ns ± 1%  12.7ns ± 2%   +0.68%  (p=0.001 n=20+19)
Parse/1-byte-16          2.23ns ± 2%  1.95ns ± 2%  -12.66%  (p=0.000 n=19+20)
Parse/2-byte-16          2.31ns ± 2%  2.01ns ± 2%  -12.87%  (p=0.000 n=20+19)
Parse/4-byte-16          2.24ns ± 2%  2.21ns ± 3%   -1.26%  (p=0.020 n=20+20)
Parse/8-byte-16          2.56ns ± 1%  2.74ns ± 2%   +7.16%  (p=0.000 n=19+19)
Append/1-byte-16         2.55ns ± 2%  2.54ns ± 3%     ~     (p=0.101 n=20+20)
Append/2-byte-16         3.67ns ± 2%  3.64ns ± 1%   -0.78%  (p=0.002 n=20+20)
Append/4-byte-16         3.62ns ± 2%  3.62ns ± 2%     ~     (p=0.931 n=20+20)
Append/8-byte-16         3.91ns ± 1%  3.86ns ± 1%   -1.28%  (p=0.000 n=18+20)
AppendWithLen/1-byte-16  3.89ns ± 2%  3.85ns ± 1%   -1.15%  (p=0.000 n=20+20)
AppendWithLen/2-byte-16  5.31ns ± 2%  5.25ns ± 1%   -1.15%  (p=0.000 n=20+18)
AppendWithLen/4-byte-16  5.56ns ± 2%  5.49ns ± 1%   -1.24%  (p=0.000 n=20+19)
AppendWithLen/8-byte-16  5.85ns ± 1%  5.77ns ± 1%   -1.39%  (p=0.000 n=20+19)
```